@@ -8,8 +8,7 @@ extension JSONValue: Decodable {
             do {
                 self = try matcher(decoder)
                 return
-            }
-            catch DecodingError.typeMismatch { continue }
+            } catch DecodingError.typeMismatch { continue }
         }
 
         throw DecodingError.typeMismatch(JSONValue.self,
@@ -21,25 +20,25 @@ extension JSONValue: Decodable {
 extension JSONValue: Encodable {
     public func encode(to encoder: Encoder) throws {
         switch self {
-        case .string(let string):
+        case let .string(string):
             var container = encoder.singleValueContainer()
             try container.encode(string)
 
         case .number:
             var container = encoder.singleValueContainer()
-            try container.encode(try self.decimalValue())
+            try container.encode(decimalValue())
 
-        case .bool(let value):
+        case let .bool(value):
             var container = encoder.singleValueContainer()
             try container.encode(value)
 
-        case .object(keyValues: let keyValues):
+        case let .object(keyValues: keyValues):
             var container = encoder.container(keyedBy: StringKey.self)
             for (key, value) in keyValues {
                 try container.encode(value, forKey: StringKey(key))
             }
 
-        case .array(let values):
+        case let .array(values):
             var container = encoder.unkeyedContainer()
             for value in values {
                 try container.encode(value)
@@ -67,7 +66,7 @@ private func decodeBool(decoder: Decoder) throws -> JSONValue {
 private func decodeObject(decoder: Decoder) throws -> JSONValue {
     let object = try decoder.container(keyedBy: StringKey.self)
     let pairs = try object.allKeys.map(\.stringValue).map { key in
-        (key, try object.decode(JSONValue.self, forKey: StringKey(key)))
+        try (key, object.decode(JSONValue.self, forKey: StringKey(key)))
     }
     return .object(keyValues: pairs)
 }
@@ -76,7 +75,9 @@ private func decodeArray(decoder: Decoder) throws -> JSONValue {
     var array = try decoder.unkeyedContainer()
     var result: [JSONValue] = []
     if let count = array.count { result.reserveCapacity(count) }
-    while !array.isAtEnd { result.append(try array.decode(JSONValue.self)) }
+    while !array.isAtEnd {
+        try result.append(array.decode(JSONValue.self))
+    }
     return .array(result)
 }
 
@@ -88,12 +89,13 @@ private func decodeNil(decoder: Decoder) throws -> JSONValue {
 }
 
 // MARK: - StringKey
-private struct StringKey: CodingKey, Hashable, CustomStringConvertible {
-    public var description: String { stringValue }
 
-    public let stringValue: String
-    public init(_ string: String) { self.stringValue = string }
-    public init?(stringValue: String) { self.init(stringValue) }
-    public var intValue: Int? { nil }
-    public init?(intValue: Int) { nil }
+private struct StringKey: CodingKey, Hashable, CustomStringConvertible {
+    var description: String { stringValue }
+
+    let stringValue: String
+    init(_ string: String) { stringValue = string }
+    init?(stringValue: String) { self.init(stringValue) }
+    var intValue: Int? { nil }
+    init?(intValue _: Int) { nil }
 }
