@@ -1,15 +1,14 @@
 //
-//  JSONCompactorTests.swift
-//  
+//  CompactorTests.swift
+//
 //
 //  Created by Rob Napier on 2/3/24.
 //
 
-import XCTest
 import RNAJSON
+import XCTest
 
 final class JSONCompactorTests: XCTestCase {
-
     func testLeadingAndTrailingWhitespaceIsSkipped() async throws {
         let json = Array(#"""
 
@@ -41,7 +40,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = #"{"id":142,"name":"aerodactyl","types":[{"type":{"name":"rock","url":"https://pokeapi.co/api/v2/type/6/"},"slot":1},{"type":{"name":"flying","url":"https://pokeapi.co/api/v2/type/3/"},"slot":2}]}"#
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -109,7 +108,7 @@ final class JSONCompactorTests: XCTestCase {
 
         let expected = ###"["JSON Test Pattern pass1",{"object with 1 member":["array with 1 element"]},{},[],-42,true,false,null,{"integer":1234567890,"real":-9876.543210,"e":0.123456789e-12,"E":1.234567890E+34,"":23456789012E66,"zero":0,"one":1,"space":" ","quote":"\"","backslash":"\\","controls":"\b\f\n\r\t","slash":"/ & \/","alpha":"abcdefghijklmnopqrstuvwyz","ALPHA":"ABCDEFGHIJKLMNOPQRSTUVWYZ","digit":"0123456789","0123456789":"digit","special":"`1~!@#$%^&*()_+-={':[,]}|;.</>?","hex":"\u0123\u4567\u89AB\uCDEF\uabcd\uef4A","true":true,"false":false,"null":null,"array":[],"object":{},"address":"50 St. James Street","url":"http://www.JSON.org/","comment":"// /* <!-- --","# -- --> */":" "," s p a c e d ":[1,2,3,4,5,6,7],"compact":[1,2,3,4,5,6,7],"jsontext":"{\"object with 1 member\":[\"array with 1 element\"]}","quotes":"&#34; \u0022 %22 0x22 034 &#x22;","\/\\\"\uCAFE\uBABE\uAB98\uFCDE\ubcda\uef4A\b\f\n\r\t`1~!@#$%^&*()_+-=[]{}|;:',./<>?":"A key can be any string"},0.5,98.6,99.44,1066,1e1,0.1e1,1e-1,1e00,2e+00,2e-00,"rosebud"]"###
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -121,7 +120,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = ###"[1,2,3]"###
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -133,7 +132,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = ###"[1,2,3]"###
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -168,7 +167,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = #"{"id":142,"name":"aerodactyl","types":[{"type":{"name":"rock","url":"https://pokeapi.co/api/v2/type/6/"},"slot":1},{"type":{"name":"flying","url":"https://pokeapi.co/api/v2/type/3/"},"slot":2}]}"#
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -180,7 +179,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = ###"[]"###
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
 
@@ -192,8 +191,7 @@ final class JSONCompactorTests: XCTestCase {
         let expected = ###"{}"###
 
         let compactor = JSONCompactor()
-        let result = String(decoding: try compactor.compact(from: json), as: UTF8.self)
+        let result = try String(decoding: compactor.compact(from: json), as: UTF8.self)
         XCTAssertEqual(result, expected)
     }
-
 }
