@@ -1,50 +1,48 @@
 //
 //  JSONPointerTests.swift
-//  
+//
 //
 //  Created by Rob Napier on 6/26/22.
 //
 
-import XCTest
-import RNAJSON
 import JSONValue
+import RNAJSON
+import XCTest
 
 final class JSONPointerTests: XCTestCase {
-
     let aerodactyl = Array("""
-{
-    "id": 142,
-    "name": "aerodactyl",
-    "types": [{
-            "type": {
-                "name": "rock",
-                "url": "https://pokeapi.co/api/v2/type/6/"
+    {
+        "id": 142,
+        "name": "aerodactyl",
+        "types": [{
+                "type": {
+                    "name": "rock",
+                    "url": "https://pokeapi.co/api/v2/type/6/"
+                },
+                "slot": 1
             },
-            "slot": 1
-        },
-        {
-            "type": {
-                "name": "flying",
-                "url": "https://pokeapi.co/api/v2/type/3/"
-            },
-            "slot": 2
-        }
-    ]
-}
-""".utf8)
+            {
+                "type": {
+                    "name": "flying",
+                    "url": "https://pokeapi.co/api/v2/type/3/"
+                },
+                "slot": 2
+            }
+        ]
+    }
+    """.utf8)
 
     func testJSONParse() async throws {
-
         let tokens = AsyncJSONTokenSequence(aerodactyl)
         let result = try await JSONValue(from: tokens)
 
         let expected = try JSONDecoder().decode(JSONValue.self, from: Data(aerodactyl))
 
         XCTAssertEqual(result, expected)
-    }    
+    }
 }
 
-//private extension XCTest {
+// private extension XCTest {
 //    // Does not check key order
 //    func XCTAssertDeepEquivalent(_ lhs: JSONValue, _ rhs: JSONValue) throws {
 //        switch (lhs, rhs) {
@@ -64,4 +62,4 @@ final class JSONPointerTests: XCTestCase {
 //            XCTAssertEqual(lhs, rhs)
 //        }
 //    }
-//}
+// }
